Refactor scripts into package tools and add entry points
