--- scripts/BACpypes.ini
@@ -1,8 +1,0 @@
-[BACpypes]
-objectName: Betelgeuse
-address: 10.0.2.15/24 
-objectIdentifier: 599
-maxApduLengthAccepted: 1024
-segmentationSupported: segmentedBoth
-vendorIdentifier: 15
-
